AI_service/requirements.txt
fastapi
uvicorn[standard]
youtube-transcript-api
yt-dlp
sentence-transformers
faiss-cpu
huggingface-hub
reportlab
python-dotenv
httpx
pydantic
numpy
